Jackson serialization for facets

+++ main/src/com/google/refine/util/SerializationFilters.java
@@ -66,8 +66,8 @@ public class SerializationFilters {
         @Override
         public void serialize(Double arg0, JsonGenerator gen, SerializerProvider s)
                 throws IOException {
-            if (new Double(arg0.intValue()).equals(arg0)) {
-                gen.writeNumber(arg0.intValue());
+            if (new Double(arg0.longValue()).equals(arg0)) {
+                gen.writeNumber(arg0.longValue());
             } else {
                 gen.writeNumber(arg0);
             }
